fix(csa): allow git var and hook-local git -C probes (#3069)
identity preflight `git var GIT_AUTHOR_IDENT`/`GIT_COMMITTER_IDENT` and hook-local `git -C` probes are allowed; real remote push stays denied.
